+++ movies/movies.go
@@ -53,14 +53,14 @@ func (m *Movie) Add(scene *scenes.Scene) {
 }
 
 // Draw draws the movie
-func (m *Movie) Draw(scale float32) {
+func (m *Movie) Draw(scale int) {
 	if m.currentScene != nil {
 		m.currentScene.Draw(scale)
 	}
 }
 
 // Update updates the movie
-func (m *Movie) Update(scale float32) (err error) {
+func (m *Movie) Update(scale int) (err error) {
 	if m.currentScene != nil {
 		err = m.currentScene.Update(scale)
 	}
